[Feat][Attention] Add RoPE to dense GQA
