+++ src/tileops/kernels/__init__.py
@@ -3,7 +3,8 @@ from .attention import (
     GQABwdWgmmaPipelinedKernel,
     GQADecodeKernel,
     GQADecodePagedKernel,
-    GQADensePrefillCausalWsKernel,
+    GQADenseCausalWsKernel,
+    GQADenseSlidingWindowKernel,
     GQAFwdFP8Fa3ContractPtxAccBN224WsTmaVKernel,
     GQAFwdWgmmaPipelinedKernel,
     GQAPrefillFwdKernel,
@@ -11,7 +12,6 @@ from .attention import (
     GQAPrefillPagedWithKVCacheFwdKernel,
     GQAPrefillPagedWithKVCacheRopeAppendKernel,
     GQAPrefillPagedWithKVCacheRopeFwdKernel,
-    GQASlidingWindowFwdWgmmaPipelinedKernel,
     GQASlidingWindowVarlenFwdWgmmaPipelinedKernel,
     MHADecodeKernel,
     MHADecodePagedKernel,
@@ -146,7 +146,8 @@ __all__ = [
     "GQABwdWgmmaPipelinedKernel",
     "GQADecodeKernel",
     "GQADecodePagedKernel",
-    "GQADensePrefillCausalWsKernel",
+    "GQADenseCausalWsKernel",
+    "GQADenseSlidingWindowKernel",
     "GQAFwdFP8Fa3ContractPtxAccBN224WsTmaVKernel",
     "GQAFwdWgmmaPipelinedKernel",
     "GQAPrefillFwdKernel",
@@ -154,7 +155,6 @@ __all__ = [
     "GQAPrefillPagedWithKVCacheFwdKernel",
     "GQAPrefillPagedWithKVCacheRopeAppendKernel",
     "GQAPrefillPagedWithKVCacheRopeFwdKernel",
-    "GQASlidingWindowFwdWgmmaPipelinedKernel",
     "GQASlidingWindowVarlenFwdWgmmaPipelinedKernel",
     "GatedDeltaNetBwdKernel",
     "GatedDeltaNetDecodeFP32Kernel",
