+++ tests/ops/attention/test_gqa.py
@@ -7,15 +7,15 @@ from torch.nn.attention import SDPBackend, sdpa_kernel
 
 from tests.test_base import FixtureBase, TestBase
 from tileops.kernels.attention import (
-    GQADensePrefillCausalWsKernel,
-    GQASlidingWindowFwdWgmmaPipelinedKernel,
+    GQADenseCausalWsKernel,
+    GQADenseSlidingWindowKernel,
 )
 from tileops.ops import (
     GroupedQueryAttentionBwdOp,
     GroupedQueryAttentionDenseFwdOp,
     GroupedQueryAttentionPrefillVarlenFwdOp,
 )
-from tileops.utils import is_h200
+from tileops.utils import get_sm_version
 from workloads.attention.gqa import GroupedQueryAttentionBwdWorkload
 
 
@@ -83,31 +83,97 @@ def _gqa_prefill_ref(
     return output.transpose(1, 2).to(q.dtype).contiguous()
 
 
+def _apply_dense_rope(
+    x: torch.Tensor,
+    positions: torch.Tensor,
+    cos: torch.Tensor,
+    sin: torch.Tensor,
+    *,
+    rotary_dim: int,
+    layout: str,
+) -> torch.Tensor:
+    half = rotary_dim // 2
+    x_rot = x[..., :rotary_dim].float()
+    c = cos[positions].view(1, x.shape[1], 1, half).float()
+    s = sin[positions].view(1, x.shape[1], 1, half).float()
+    if layout == "neox":
+        x0, x1 = x_rot[..., :half], x_rot[..., half:]
+    else:
+        x0, x1 = x_rot[..., 0::2], x_rot[..., 1::2]
+    y0, y1 = x0 * c - x1 * s, x1 * c + x0 * s
+    rotated = (
+        torch.cat((y0, y1), dim=-1)
+        if layout == "neox"
+        else torch.stack((y0, y1), dim=-1).flatten(-2)
+    )
+    return torch.cat((rotated.to(x.dtype), x[..., rotary_dim:]), dim=-1).contiguous()
+
+
+@pytest.mark.parametrize(
+    "rope_layout, rotary_dim, dtype",
+    [
+        (None, None, torch.float16),
+        ("neox", 64, torch.float16),
+        ("interleaved", 64, torch.float16),
+        ("neox", None, torch.bfloat16),
+    ],
+)
 @pytest.mark.smoke
-def test_gqa_dense_h200_main_kernel_matches_reference() -> None:
-    if not is_h200():
-        pytest.skip("Dense warp-specialized prefill is currently supported on H200")
-    batch, seq_len_q, seq_len_kv, heads, heads_kv, dim = 1, 96, 150, 8, 2, 128
-    q = torch.randn(batch, seq_len_q, heads, dim, device="cuda", dtype=torch.float16)
-    k = torch.randn(batch, seq_len_kv, heads_kv, dim, device="cuda", dtype=torch.float16)
+def test_gqa_dense_sm90_main_kernel_matches_reference(
+    rope_layout: Optional[str], rotary_dim: Optional[int], dtype: torch.dtype
+) -> None:
+    if not torch.cuda.is_available() or get_sm_version() != 90:
+        pytest.skip("Dense warp-specialized prefill requires SM90")
+    batch, seq_len_q, seq_len_kv, heads, heads_kv, dim = 1, 160, 270, 8, 2, 128
+    q = torch.randn(batch, seq_len_q, heads, dim, device="cuda", dtype=dtype)
+    k = torch.randn(batch, seq_len_kv, heads_kv, dim, device="cuda", dtype=dtype)
     v = torch.randn_like(k)
 
-    op = GroupedQueryAttentionDenseFwdOp()
-    output = op(q, k, v)
+    if rope_layout is None:
+        op = GroupedQueryAttentionDenseFwdOp()
+        output = op(q, k, v)
+        q_ref, k_ref = q, k
+    else:
+        resolved_rotary_dim = dim if rotary_dim is None else rotary_dim
+        angles = torch.randn(seq_len_kv, resolved_rotary_dim // 2, device="cuda") * 0.1
+        rope_cos, rope_sin = angles.cos().to(dtype), angles.sin().to(dtype)
+        op = GroupedQueryAttentionDenseFwdOp(
+            pos_encoding_mode="rope", rotary_dim=rotary_dim, rope_layout=rope_layout
+        )
+        output = op(q, k, v, rope_cos=rope_cos, rope_sin=rope_sin)
+        q_positions = torch.arange(seq_len_kv - seq_len_q, seq_len_kv, device="cuda")
+        k_positions = torch.arange(seq_len_kv, device="cuda")
+        q_ref = _apply_dense_rope(
+            q,
+            q_positions,
+            rope_cos,
+            rope_sin,
+            rotary_dim=resolved_rotary_dim,
+            layout=rope_layout,
+        )
+        k_ref = _apply_dense_rope(
+            k,
+            k_positions,
+            rope_cos,
+            rope_sin,
+            rotary_dim=resolved_rotary_dim,
+            layout=rope_layout,
+        )
 
     torch.testing.assert_close(
         output,
-        _gqa_prefill_ref(q, k, v, heads=heads, heads_kv=heads_kv, is_causal=True),
-        atol=5e-3,
+        _gqa_prefill_ref(q_ref, k_ref, v, heads=heads, heads_kv=heads_kv, is_causal=True),
+        atol=2e-2 if dtype == torch.bfloat16 else 5e-3,
         rtol=1e-5,
     )
-    assert isinstance(next(iter(op.iter_kernels())), GQADensePrefillCausalWsKernel)
+    assert isinstance(next(iter(op.iter_kernels())), GQADenseCausalWsKernel)
 
 
+@pytest.mark.parametrize("use_rope", [False, True])
 @pytest.mark.smoke
-def test_gqa_dense_h200_sliding_window_kernel_matches_reference() -> None:
-    if not is_h200():
-        pytest.skip("Dense sliding-window prefill is currently supported on H200")
+def test_gqa_dense_sm90_sliding_window_kernel_matches_reference(use_rope: bool) -> None:
+    if not torch.cuda.is_available() or get_sm_version() != 90:
+        pytest.skip("Dense sliding-window prefill requires SM90")
     batch, seq_len, heads, heads_kv, dim = 1, 256, 8, 2, 128
     window_size_left = 64
     sm_scale = 0.125
@@ -116,13 +182,36 @@ def test_gqa_dense_h200_sliding_window_kernel_matches_reference() -> None:
     k = torch.randn(batch, seq_len, heads_kv, dim, device="cuda", dtype=torch.float16)
     v = torch.randn_like(k)
 
+    rotary_dim = 64
+    if use_rope:
+        angles = torch.randn(seq_len, rotary_dim // 2, device="cuda") * 0.1
+        rope_cos, rope_sin = angles.cos().to(q.dtype), angles.sin().to(q.dtype)
+    else:
+        rope_cos = rope_sin = None
     op = GroupedQueryAttentionDenseFwdOp(
         window_size_left=window_size_left,
         window_size_right=0,
         sm_scale=sm_scale,
         softcap=softcap,
+        pos_encoding_mode="rope" if use_rope else "none",
+        rotary_dim=rotary_dim if use_rope else None,
     )
-    output = op(q, k, v)
+    output = op(
+        q,
+        k,
+        v,
+        rope_cos=rope_cos,
+        rope_sin=rope_sin,
+    )
+    if use_rope:
+        assert rope_cos is not None and rope_sin is not None
+        positions = torch.arange(seq_len, device="cuda")
+        q = _apply_dense_rope(
+            q, positions, rope_cos, rope_sin, rotary_dim=rotary_dim, layout="neox"
+        )
+        k = _apply_dense_rope(
+            k, positions, rope_cos, rope_sin, rotary_dim=rotary_dim, layout="neox"
+        )
 
     torch.testing.assert_close(
         output,
@@ -141,7 +230,7 @@ def test_gqa_dense_h200_sliding_window_kernel_matches_reference() -> None:
         atol=5e-3,
         rtol=1e-5,
     )
-    assert isinstance(next(iter(op.iter_kernels())), GQASlidingWindowFwdWgmmaPipelinedKernel)
+    assert isinstance(next(iter(op.iter_kernels())), GQADenseSlidingWindowKernel)
 
 
 class GroupedQueryAttentionBwdFixture(FixtureBase):
